Update syno_enable_m2_volume.sh
v1.1.11
- Now only downloads bc if not in PATH or in script location.
- Now asks to download bc (if --autoupdate option not used).

+++ syno_enable_m2_volume.sh
@@ -5,13 +5,13 @@
 #
 # Github: https://github.com/007revad/Synology_enable_M2_volume
 # Script verified at https://www.shellcheck.net/
-# Tested on DSM 7.2
+# Tested on DSM 7.2 and 7.2.1
 #
 # To run in a shell (replace /volume1/scripts/ with path to script):
 # sudo /volume1/scripts/syno_enable_m2_volume.sh
 #------------------------------------------------------------------------------
 
-scriptver="v1.0.10"
+scriptver="v1.1.11"
 script=Synology_enable_M2_volume
 repo="007revad/Synology_enable_M2_volume"
 scriptname=syno_enable_m2_volume
@@ -178,6 +178,13 @@ if [[ $buildphase == GM ]]; then buildphase=""; fi
 if [[ $smallfixnumber -gt "0" ]]; then smallfix="-$smallfixnumber"; fi
 echo -e "$model DSM $productversion-$buildnumber$smallfix $buildphase\n"
 
+
+# Get StorageManager version
+storagemgrver=$(synopkg version StorageManager)
+# Show StorageManager version
+if [[ $storagemgrver ]]; then echo -e "StorageManager $storagemgrver\n"; fi
+
+
 # Show options used
 echo "Using options: ${args[*]}"
 
@@ -186,23 +193,48 @@ echo "Using options: ${args[*]}"
 # Install bc command if missing
 
 if ! which bc >/dev/null ; then
-    # Download bc
-    echo -e "\nDownloading bc"
-    curl -kL "https://raw.githubusercontent.com/${repo}/main/bin/bc" -o /tmp/bc
+    if [[ -f ./dtb/${modelname}_model.dtb ]]; then
+        # bc exists in bin folder in script location
+        bcfile="./bin/bc"
+    elif [[ -f ./${modelname}_model.dtb ]]; then
+        # bc exists in same folder as script
+        bcfile="./bc"
+    else
+        # Download bc
+        if [[ $autoupdate == "yes" ]]; then
+            reply=y
+        else
+            echo "Need to download bc."
+            echo -e "${Cyan}Do you want to download bc?${Off} [y/n]"
+            read -r -t 30 reply
+        fi
+
+        if [[ ${reply,,} == "y" ]]; then
+            echo -e "\nDownloading bc"
+            curl -kL "https://raw.githubusercontent.com/${repo}/main/bin/bc" -o /tmp/bc
+            bcfile="/tmp/bc"
+        else
+            echo "Cannot run without bc!"
+            exit 1
+        if
+    fi
 
     # Set bc executable
-    chmod a+x /tmp/bc
+    #chmod a+x /tmp/bc
+    chmod a+x "$bcfile"
 
     # Copy bc to /usr/bin
-    cp -p /tmp/bc /usr/bin/bc
+    #cp -p /tmp/bc /usr/bin/bc
+    cp -p "$bcfile" /usr/bin/bc
+fi
 
-    # Check bc is executable
-    if [[ ! -x /usr/bin/bc ]]; then
-        ding
-        echo "/usr/bin/bc not found or not executable!"
-        exit 1
-    fi
-    echo
+# Check bc is executable
+bcfile=$(which bc)
+if [[ ! -x "$bcfile" ]]; then
+    ding
+    echo "bc not found or not executable!"
+    echo "bc: $bcfile"
+    exit 1
 fi
 
 
@@ -408,8 +440,11 @@ rebootmsg(){
     read -r -t 10 answer
     if [[ ${answer,,} != "yes" ]]; then exit; fi
 
-    # Reboot in the background so user can see DSM's "going down" message
-    reboot &
+#    # Reboot in the background so user can see DSM's "going down" message
+#    reboot &
+    [ -x /usr/syno/sbin/synopoweroff ] &&
+    /usr/syno/sbin/synopoweroff -r ||
+    reboot
 }
